build: remove enableStrongSkippingMode = true
Enable strong skipping.
Strong Skipping is a mode that improves the runtime performance of your application by skipping unnecessary invocations of composable functions for which the parameters have not changed. In particular, when enabled, Composable functions with unstable parameters become skippable and lambdas with unstable captures will be memoized.
For more information, see this link:
Strong skipping
This feature is enabled by default. To disable, provide this feature flag in a disabled state: